+++ app/src/main/java/com/example/myapplication/data/model/entity/UpcomingKey.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "remote_key_upcoming")
 data class UpcomingKey(
     @PrimaryKey(autoGenerate = true)
-    val upcomingId: Int,
-    val movieId: Int,
+    val upcomingId: Long,
+    val movieId: Long,
     val page: Int?
 )
