+++ app/src/main/java/com/example/myapplication/data/model/pos/UpcomingPos.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "upcoming_pos")
 data class UpcomingPos(
     @PrimaryKey(autoGenerate = true)
-    val upcomingId: Int,
-    val movieId: Int,
+    val upcomingId: Long,
+    val movieId: Long,
     val position: Int
 )
